Topic/2024.1/linker performance optimizations 2

+++ Metalama.Framework.Engine/CodeModel/References/RefEqualityComparer{T}.cs
@@ -62,7 +62,8 @@ namespace Metalama.Framework.Engine.CodeModel.References
                     ? this.StructuralSymbolComparer.GetHashCode( xSymbol )
                     : RuntimeHelpers.GetHashCode( obj.Target );
 
-                return HashCode.Combine( targetHashCode, obj.TargetKind );
+                // PERF: Cast enum to int otherwise it will be boxed on .NET Framework.
+                return HashCode.Combine( targetHashCode, (int)obj.TargetKind );
             }
         }
